Vladimir_Kovtunovskiy/homework5/voice_data_agent/requirements.txt
google-generativeai
google-adk